Fix hIST chunk position to match W3C specification
The hIST chunk was incorrectly positioned in the code. According to the
W3C specification, the hIST chunk must appear before the first IDAT
chunk and after PLTE (if present). This commit corrects the chunk
positioning to comply with the specification.

This is a cherry-pick of commit 9ecda6eaf5726e0248cf60ab8926423c476a66c0
from branch 'libpng18'.

Signed-off-by: Chris Lilley <chris@w3.org>
Signed-off-by: Cosmin Truta <ctruta@gmail.com>